+++ .trebired/logger/config.ts
@@ -1,0 +1,4 @@
+export default {
+  forVersion: "2.7.0",
+  prefix: "trebired.grammar",
+};